[#4210] ApiObjectRef.setObject value handling enhancement
If the ApiObjectRef.setObject is called with the current object, then it
is possible that an invalid PropertyChange will be fired, if in the
meantime there was a modification through a wrapper, and after that the
object was modified directly. In this case the already existing
PropertyChange.oldValue should be used as a check for modification, not
the object's value.