[#11249] componentApi initial implementation
As we move towards a more generic component / page handling, we need
generic services for the client. This initial implementation consist of:
- load (getComponentModel)
- execute (performAction)
- download (downloadItem, with previous impl. still working)
It also introduces some very important concepts, like ComponentModel
(which is the data centric generic UI bean), ViewContextChange (which is
a result of an executed action, and can be used to restore client state
without furhter server calls).