[#11305] fix lookup based on URI comparison
When users are deleted, `OrgApiStorageImpl` failed to remove them from the object-map of active users: it relied on `getKeyByValue(URI, ObjectMap)` which tried to compare `URI`s with no regard to the fact that two values are not equal if they differ only in version numbers.

With this fix, `removeUser(URI)` will correctly remove a user from the active users, and allow˛`restoreDeletedUser(URI)` to correctly restore inactive users (if their username is still available when the restoration is attempted).