[#13913] ensure invocations can proceed when an authenticated session for invocation channels is not readily available
asynchronous invocation channels now interpret the technical username applications may configure for them. If they are present, a session authenticated with this user is employed to execute asynchronous invocation requests (for in these threads the per-thread SecurityContext of the invoker is not available, even if that thread was considered authenticated).