[#21514] implement ReadCache in ObjectApi
Cache is handled in thread local, and will be disabled in first save
with a warning, or with an explicit disable call.
Latest and exact version objectNodes are cached, as well as lastModified
and exists infos. Only ObjectApi methods are involved, lower level
storage calls will read objects.
For now, only SearchIndexImpl uses it, and be default this functionality
is turned off. You can turn it on in application.properties
smartbit4all.objectapi.useReadCache=true